fixed errors in guests, requests and staff

+++ app.py
@@ -580,7 +580,7 @@ def check_staff(staff_id):
 @app.route('/delete/<int:guest_id>')
 def delete_guest(guest_id):
     cursor = mysql.connection.cursor()
-    cursor.execute("DELETE FROM Guest WHERE guest_id = %s", (guest_id,))
+    cursor.execute("DELETE FROM guest WHERE guest_id = %s", (guest_id,))
     mysql.connection.commit()
     cursor.close()
     return redirect('/guests')
@@ -597,7 +597,7 @@ def editGuest(guest_id):
         phone = request.form['phone']
 
         cursor.execute("""
-            UPDATE Guest SET first_name = %s, middle_name = %s,
+            UPDATE guest SET first_name = %s, middle_name = %s,
             last_name = %s, email = %s, phone = %s
             WHERE guest_id = %s
         """, (first_name, middle_name, last_name, email, phone, guest_id))
@@ -606,7 +606,7 @@ def editGuest(guest_id):
         return redirect('/guests')
 
     #GET method - show the edit form
-    cursor.execute("SELECT * FROM Guest WHERE guest_id = %s", (guest_id,))
+    cursor.execute("SELECT * FROM guest WHERE guest_id = %s", (guest_id,))
     guest = cursor.fetchone()
     cursor.close()
 
@@ -623,6 +623,7 @@ def show_requests():
     cursor.execute("""
         SELECT r.request_id, r.booking_id, r.quantity, r.unit_cost, r.total_cost,
                r.status, r.request_time, r.staff_id, 
+               r.last_update, r.timestamp,
                s.name AS service_name, f.name AS food_name,
                s.category as service_category, f.category as food_category,
                s.type as service_type, f.type as food_type,
@@ -975,6 +976,7 @@ def add_guests():
     #get new data and save logs
     if new_guest_id:
         new_data_for_log = {
+            'guest_id': new_guest_id,
             'first_name': first_name,
             'middle_name': middle_name,
             'last_name': last_name,
@@ -1064,6 +1066,8 @@ def add_request():
     total_cost = quantity * unit_cost
     notes = request.form['notes']
     status = request.form['status']
+    last_update = session['username']
+    timestamp = datetime.now()
 
       # Get the name value and parse it
     name_value = request.form.get('name')  # e.g., "service_11" or "food_5"
@@ -1088,11 +1092,42 @@ def add_request():
     room_number = row['room_number'] if row else None
 
     cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+    new_request_id = None #change 
+    
     cursor.execute("""
-    INSERT INTO requests (booking_id, guest_id, service_id, item_id, quantity, unit_cost, total_cost, status, request_time, room_number, notes)
-    VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
-    """, (booking_id, guest_id, service_id, item_id, quantity, unit_cost, total_cost, status, request_time, room_number, notes))
+    INSERT INTO requests (booking_id, guest_id, service_id, item_id, quantity, unit_cost, total_cost, status, request_time, room_number, notes, last_update, timestamp)
+    VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
+    """, (booking_id, guest_id, service_id, item_id, quantity, unit_cost, total_cost, status, request_time, room_number, notes, last_update, timestamp))
+    new_request_id = cursor.lastrowid #Get the ID of the newly inserted record; change room_id
     mysql.connection.commit()
+    
+    #Get new data and save logs
+    if new_request_id:  #change room_id
+        new_data_for_log = {
+            'request_id': new_request_id,
+            'booking_id': booking_id,  #change field names
+            'guest_id': guest_id,
+            'service_id': service_id,
+            'item_id': item_id,
+            'quantity': quantity,
+            'unit_cost': unit_cost,
+            'total_cost': total_cost,
+            'status': status,
+            'request_time': request_time,
+            'room_number': room_number,
+            'notes': notes,
+            'last_update': last_update,
+            'timestamp': timestamp
+        }
+        log_audit_event(
+            actor_id = session['username'],
+            timestamp=timestamp,
+            table_name='requests',  #change 
+            action_type='INSERT', 
+            record_id=str(new_request_id),  #change
+            old_data=None,           # Record did not exist, so old_data is None
+            new_data=new_data_for_log 
+        )
     cursor.close()
 
     return redirect('/requests')
@@ -1112,6 +1147,8 @@ def update_request():
     unit_cost = float(request.form['edit_unit_cost'])
     total_cost = quantity * unit_cost
     notes = request.form['edit_notes']
+    last_update = session['username']
+    timestamp = datetime.now()
 
       # Get the name value and parse it
     name_value = request.form.get('edit_name')  # e.g., "service_11" or "food_5"
@@ -1128,6 +1165,29 @@ def update_request():
                 service_id = None
 
     cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+    
+    #Get old data
+    cursor.execute("SELECT * FROM requests WHERE request_id = %s", (request_id,)) #change table and field names
+    old_data = cursor.fetchone() 
+
+    if not old_data:
+        # Handle error: record not found
+        return "Request not found", 404 #change message
+    
+    #Get new data 
+    new_data = old_data.copy()
+    new_data['request_id'] = request_id  #change ALL field names (should be similar to the table)
+    new_data['booking_id'] = booking_id
+    new_data['service_id'] = service_id
+    new_data['item_id'] = item_id
+    new_data['quantity'] = quantity
+    new_data['status'] = status
+    new_data['request_time'] = request_time
+    new_data['unit_cost'] = unit_cost
+    new_data['total_cost'] = total_cost
+    new_data['notes'] = notes
+    new_data['last_update'] = last_update
+    new_data['timestamp'] = timestamp
 
     cursor.execute("""
     UPDATE requests
@@ -1139,11 +1199,24 @@ def update_request():
         total_cost = %s,
         status = %s,
         request_time = %s,
-        notes = %s
+        notes = %s,
+        last_update = %s,
+        timestamp = %s
     WHERE request_id = %s
-""", (booking_id, service_id, item_id, quantity, unit_cost, total_cost, status, request_time, notes, request_id))
+""", (booking_id, service_id, item_id, quantity, unit_cost, total_cost, status, request_time, notes, last_update, timestamp, request_id))
 
     mysql.connection.commit()
+    
+    #Save logs
+    log_audit_event(
+        actor_id=session['username'],
+        timestamp=timestamp,
+        table_name='requests',  #change table name
+        action_type='UPDATE', 
+        record_id=str(request_id), #change field name
+        old_data=old_data, 
+        new_data=new_data 
+    )
     cursor.close()
 
     return redirect('/requests')
@@ -1154,7 +1227,30 @@ def deleteRequest(request_id):
     #Delete assignments first
     #cursor.execute("DELETE FROM StaffAssignments WHERE request_id = %s", (request_id,))
     #Then delete the request
+    #Get old data
+    cursor.execute("SELECT * FROM requests WHERE request_id =%s", (request_id,)) #change table and field name
+    old_data = cursor.fetchone()
+    timestamp = datetime.now()
+    
+    if not old_data:
+        cursor.close()
+        #Handle case where the room ID doesn't exist
+        return "Request not found or already deleted", 404 #change message
+    
+    
     cursor.execute("DELETE FROM requests WHERE request_id = %s", (request_id,))
+    
+    #Save logs
+    log_audit_event(
+        actor_id = session['username'],
+        timestamp=timestamp,
+        table_name='requests',  #change
+        action_type='DELETE', 
+        record_id=str(request_id), #change
+        old_data=old_data, 
+        new_data=None 
+    )
+    
     mysql.connection.commit()
     cursor.close()
     return redirect('/requests')
@@ -1544,12 +1640,12 @@ def add_massage():
     #Get new data and save logs
     if new_service_id:  #change room_id
         new_data_for_log = {
+            'service_id': new_service_id,
             'category': category,  #change field names
             'name': name,
             'description': description,
             'price': price,
             'type': type,
-            'sevice_id': new_service_id,
             'last_update': last_update,
             'timestamp': timestamp
         }
@@ -2085,7 +2181,7 @@ def add_room_guest():
 
     cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
     cursor.execute(
-        "INSERT INTO RoomGuest (room_id, guest_id, checkin_date, checkout_date) VALUES (%s, %s, %s, %s)",
+        "INSERT INTO roomguest (room_id, guest_id, checkin_date, checkout_date) VALUES (%s, %s, %s, %s)",
         (room_id, guest_id, checkin_date, checkout_date)
     )
     mysql.connection.commit()
@@ -2397,7 +2493,7 @@ def checkout():
     cursor.execute("SELECT room_id FROM bookings WHERE booking_id = %s", (booking_id,))
     room = cursor.fetchone()
     if room:
-        cursor.execute("UPDATE Room SET room_status = 'Vacant' WHERE room_id = %s", (room['room_id'],))
+        cursor.execute("UPDATE room SET room_status = 'Vacant' WHERE room_id = %s", (room['room_id'],))
     else:
         return "Room not found", 404
